+++ crates/teloxide-core/src/types/update.rs
@@ -29,7 +29,7 @@ pub struct Update {
 }
 
 impl Update {
-    // FIXME: rename user => from, add mentioned_users -> impl Iterator<&User>
+    // FIXME: add mentioned_users -> impl Iterator<&User>
 
     /// Returns the user that performed the action that caused this update, if
     /// known.
@@ -37,7 +37,7 @@ impl Update {
     /// This is generally the `from` field (except for `PollAnswer` where it's
     /// `user` and `Poll` with `Error` which don't have such field at all).
     #[must_use]
-    pub fn user(&self) -> Option<&User> {
+    pub fn from(&self) -> Option<&User> {
         use UpdateKind::*;
 
         let from = match &self.kind {
@@ -57,6 +57,48 @@ impl Update {
         };
 
         Some(from)
+    }
+
+    /// Returns all users that are "contained" in this `Update` structure.
+    ///
+    /// This might be useful to track information about users.
+    ///
+    /// Note that this function may return quite a few users as it scans
+    /// replies, pinned messages, message entities, "via bot" fields and more.
+    /// Also note that this function can return duplicate users.
+    pub fn mentioned_users(&self) -> impl Iterator<Item = &User> {
+        use either::Either::{Left, Right};
+        use std::iter::{empty, once};
+
+        let i0 = Left;
+        let i1 = |x| Right(Left(x));
+        let i2 = |x| Right(Right(Left(x)));
+        let i3 = |x| Right(Right(Right(Left(x))));
+        let i4 = |x| Right(Right(Right(Right(Left(x)))));
+        let i5 = |x| Right(Right(Right(Right(Right(Left(x))))));
+        let i6 = |x| Right(Right(Right(Right(Right(Right(x))))));
+
+        match &self.kind {
+            UpdateKind::Message(message)
+            | UpdateKind::EditedMessage(message)
+            | UpdateKind::ChannelPost(message)
+            | UpdateKind::EditedChannelPost(message) => i0(message.mentioned_users()),
+
+            UpdateKind::InlineQuery(query) => i1(once(&query.from)),
+            UpdateKind::ChosenInlineResult(query) => i1(once(&query.from)),
+            UpdateKind::CallbackQuery(query) => i2(query.mentioned_users()),
+            UpdateKind::ShippingQuery(query) => i1(once(&query.from)),
+            UpdateKind::PreCheckoutQuery(query) => i1(once(&query.from)),
+            UpdateKind::Poll(poll) => i3(poll.mentioned_users()),
+
+            UpdateKind::PollAnswer(answer) => i1(once(&answer.user)),
+
+            UpdateKind::MyChatMember(member) | UpdateKind::ChatMember(member) => {
+                i4(member.mentioned_users())
+            }
+            UpdateKind::ChatJoinRequest(request) => i5(request.mentioned_users()),
+            UpdateKind::Error(_) => i6(empty()),
+        }
     }
 
     /// Returns the chat in which is update has happened, if any.
@@ -81,6 +123,11 @@ impl Update {
         };
 
         Some(chat)
+    }
+
+    #[deprecated(note = "renamed to `from`", since = "0.10.0")]
+    pub fn user(&self) -> Option<&User> {
+        self.from()
     }
 }
 
@@ -156,7 +203,10 @@ pub enum UpdateKind {
     /// An error that happened during deserialization.
     ///
     /// This allows `teloxide` to continue working even if telegram adds a new
-    /// kind of updates.
+    /// kinds of updates.
+    ///
+    /// **Note that deserialize implementation always returns an empty value**,
+    /// teloxide fills in the data when doing deserialization.
     Error(Value),
 }
 
@@ -182,94 +232,63 @@ impl<'de> Deserialize<'de> for UpdateKind {
 
                 // Try to deserialize a borrowed-str key, or else try deserializing an owned
                 // string key
-                let k = map.next_key::<&str>().or_else(|_| {
+                let key = map.next_key::<&str>().or_else(|_| {
                     map.next_key::<String>().map(|k| {
                         tmp = k;
                         tmp.as_deref()
                     })
                 });
 
-                if let Ok(Some(k)) = k {
-                    let res = match k {
-                        "message" => {
-                            map.next_value::<Message>().map(UpdateKind::Message).map_err(|_| false)
+                let this = key
+                    .ok()
+                    .flatten()
+                    .and_then(|key| match key {
+                        "message" => map.next_value::<Message>().ok().map(UpdateKind::Message),
+                        "edited_message" => {
+                            map.next_value::<Message>().ok().map(UpdateKind::EditedMessage)
                         }
-                        "edited_message" => map
-                            .next_value::<Message>()
-                            .map(UpdateKind::EditedMessage)
-                            .map_err(|_| false),
-                        "channel_post" => map
-                            .next_value::<Message>()
-                            .map(UpdateKind::ChannelPost)
-                            .map_err(|_| false),
-                        "edited_channel_post" => map
-                            .next_value::<Message>()
-                            .map(UpdateKind::EditedChannelPost)
-                            .map_err(|_| false),
-                        "inline_query" => map
-                            .next_value::<InlineQuery>()
-                            .map(UpdateKind::InlineQuery)
-                            .map_err(|_| false),
+                        "channel_post" => {
+                            map.next_value::<Message>().ok().map(UpdateKind::ChannelPost)
+                        }
+                        "edited_channel_post" => {
+                            map.next_value::<Message>().ok().map(UpdateKind::EditedChannelPost)
+                        }
+                        "inline_query" => {
+                            map.next_value::<InlineQuery>().ok().map(UpdateKind::InlineQuery)
+                        }
                         "chosen_inline_result" => map
                             .next_value::<ChosenInlineResult>()
-                            .map(UpdateKind::ChosenInlineResult)
-                            .map_err(|_| false),
-                        "callback_query" => map
-                            .next_value::<CallbackQuery>()
-                            .map(UpdateKind::CallbackQuery)
-                            .map_err(|_| false),
-                        "shipping_query" => map
-                            .next_value::<ShippingQuery>()
-                            .map(UpdateKind::ShippingQuery)
-                            .map_err(|_| false),
+                            .ok()
+                            .map(UpdateKind::ChosenInlineResult),
+                        "callback_query" => {
+                            map.next_value::<CallbackQuery>().ok().map(UpdateKind::CallbackQuery)
+                        }
+                        "shipping_query" => {
+                            map.next_value::<ShippingQuery>().ok().map(UpdateKind::ShippingQuery)
+                        }
                         "pre_checkout_query" => map
                             .next_value::<PreCheckoutQuery>()
-                            .map(UpdateKind::PreCheckoutQuery)
-                            .map_err(|_| false),
-                        "poll" => map.next_value::<Poll>().map(UpdateKind::Poll).map_err(|_| false),
-                        "poll_answer" => map
-                            .next_value::<PollAnswer>()
-                            .map(UpdateKind::PollAnswer)
-                            .map_err(|_| false),
-                        "my_chat_member" => map
-                            .next_value::<ChatMemberUpdated>()
-                            .map(UpdateKind::MyChatMember)
-                            .map_err(|_| false),
-                        "chat_member" => map
-                            .next_value::<ChatMemberUpdated>()
-                            .map(UpdateKind::ChatMember)
-                            .map_err(|_| false),
+                            .ok()
+                            .map(UpdateKind::PreCheckoutQuery),
+                        "poll" => map.next_value::<Poll>().ok().map(UpdateKind::Poll),
+                        "poll_answer" => {
+                            map.next_value::<PollAnswer>().ok().map(UpdateKind::PollAnswer)
+                        }
+                        "my_chat_member" => {
+                            map.next_value::<ChatMemberUpdated>().ok().map(UpdateKind::MyChatMember)
+                        }
+                        "chat_member" => {
+                            map.next_value::<ChatMemberUpdated>().ok().map(UpdateKind::ChatMember)
+                        }
                         "chat_join_request" => map
                             .next_value::<ChatJoinRequest>()
-                            .map(UpdateKind::ChatJoinRequest)
-                            .map_err(|_| false),
+                            .ok()
+                            .map(UpdateKind::ChatJoinRequest),
+                        _ => Some(empty_error()),
+                    })
+                    .unwrap_or_else(empty_error);
 
-                        _ => Err(true),
-                    };
-
-                    let value_available = match res {
-                        Ok(ok) => return Ok(ok),
-                        Err(e) => e,
-                    };
-
-                    let mut value = serde_json::Map::new();
-                    value.insert(
-                        k.to_owned(),
-                        if value_available {
-                            map.next_value::<Value>().unwrap_or(Value::Null)
-                        } else {
-                            Value::Null
-                        },
-                    );
-
-                    while let Ok(Some((k, v))) = map.next_entry::<_, Value>() {
-                        value.insert(k, v);
-                    }
-
-                    return Ok(UpdateKind::Error(Value::Object(value)));
-                }
-
-                Ok(UpdateKind::Error(Value::Object(<_>::default())))
+                Ok(this)
             }
         }
 
@@ -317,6 +336,10 @@ impl Serialize for UpdateKind {
             UpdateKind::Error(v) => v.serialize(s),
         }
     }
+}
+
+fn empty_error() -> UpdateKind {
+    UpdateKind::Error(Value::Object(<_>::default()))
 }
 
 #[cfg(test)]