+++ crates/teloxide-core/CHANGELOG.md
@@ -7,11 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Added
+
+- `ChatPermission::can_*` helper functions ([#851][pr851])
+- `mentioned_users` functions for `CallbackQuery`, `Chat`, `ChatJoinRequest`, `ChatMemberUpdated`, `Game`, `Message`, `Poll`, `Update` which return all contained `User` instances ([#850][pr850])
+- `Message::video_chat_participants_invited` ([#850][pr850])
+- `Update::from`, a replacement for `Update::user` ([#850][pr850])
+
+[pr851]: https://github.com/teloxide/teloxide/pull/851
+
+### Deprecated
+
+- `Update::user`, use `Update::from` instead ([#850][pr850])
+
+[pr850]: https://github.com/teloxide/teloxide/pull/850
+
+## 0.9.1 - 2023-02-15
+
 ### Fixed 
 
 - `Update::user` now handles channel posts, chat member changes and chat join request updates correctly ([#835][pr835])
+- In cases when `teloxide` can't deserialize an update, error now includes the full json value ([#826][pr826])
+- Deserialization of topic messages ([#830][pr830])
 
 [pr835]: https://github.com/teloxide/teloxide/pull/835
+[pr826]: https://github.com/teloxide/teloxide/pull/826
+[pr830]: https://github.com/teloxide/teloxide/pull/830
+
+### Added
+
+- `ApiError::ImageProcessFailed` ([#825][pr825])
+
+[pr825]: https://github.com/teloxide/teloxide/pull/825
 
 ## 0.9.0 - 2023-01-17